+++ src/minter/library/ConfigIncentiveLib.sol
@@ -15,17 +15,17 @@ library ConfigIncentiveLib {
 
     /// @notice The precision at which incentive ratios are stored.
     /// @dev With decimals = 9, this gives a max ratio of 2 (200%) with precision of 0.000000001 (0.0000001%)
-    uint internal constant INCENTIVE_RATIO_DECIMALS = 9; // solhint-disable-line explicit-types
+    uint256 internal constant INCENTIVE_RATIO_DECIMALS = 9; // solhint-disable-line explicit-types
 
     /// @notice The precision at which collateral ratio bounds are stored.
     /// @dev With decimals = 6, this gives a max ratio of 4,000 (400,000%) with precision of 0.000001 (0.0001%)
-    uint internal constant COLLATERAL_RATIO_DECIMALS = 6; // solhint-disable-line explicit-types
+    uint256 internal constant COLLATERAL_RATIO_DECIMALS = 6; // solhint-disable-line explicit-types
 
     /// @notice The maximum number of fee/discount value bands that can be stored
-    uint internal constant MAX_BANDS = 8; // solhint-disable-line explicit-types
+    uint256 internal constant MAX_BANDS = 8; // solhint-disable-line explicit-types
 
     /// @notice The maximum number of collateral ratio bounds for fee/discount variation that can be stored
-    uint internal constant MAX_BOUNDS = MAX_BANDS - 1; // solhint-disable-line explicit-types
+    uint256 internal constant MAX_BOUNDS = MAX_BANDS - 1; // solhint-disable-line explicit-types
 
     ///////////////
     // Structs   //
@@ -50,8 +50,12 @@ library ConfigIncentiveLib {
     /// @notice Returns a collateral ratio bound at the given index
     function _collateralRatioUpperBounds(
         ActionIncentive memory config_,
-        uint index // solhint-disable-line explicit-types
-    ) internal pure returns (uint256 result) {
+        uint256 index // solhint-disable-line explicit-types
+    )
+        internal
+        pure
+        returns (uint256 result)
+    {
         result = (config_.slot0.decodeUint(index * 32, 32) * 10 ** (18 - COLLATERAL_RATIO_DECIMALS));
         // an upper bound of 1 ether actually means an upper bound just below 1 ether because that's where it becomes depegged
         // we treat 1 ether specially, as we can't specify 1 ether -1 so we just subtract 1 here
@@ -63,8 +67,12 @@ library ConfigIncentiveLib {
     /// @notice Returns a collateral ratio lower bound at the given index`
     function _collateralRatioLowerBounds(
         ActionIncentive memory config_,
-        uint index // solhint-disable-line explicit-types
-    ) internal pure returns (uint256 result) {
+        uint256 index // solhint-disable-line explicit-types
+    )
+        internal
+        pure
+        returns (uint256 result)
+    {
         // if we are in the lowest band, the lower bound is 0
         // else its the previous upper bound
         result = index == 0 ? 0 ether : _collateralRatioUpperBounds(config_, index - 1);
@@ -75,13 +83,13 @@ library ConfigIncentiveLib {
 
     /// @notice Returns the collateral ratio bound count
     // solhint-disable-next-line explicit-types
-    function _collateralRatioBandCount(ActionIncentive memory config_) internal pure returns (uint count) {
+    function _collateralRatioBandCount(ActionIncentive memory config_) internal pure returns (uint256 count) {
         count = config_.slot0.decodeUint(224, 8);
     }
 
     /// @notice Returns a incentive ratio at the given index
     // solhint-disable-next-line explicit-types
-    function _incentiveRatio(ActionIncentive memory config_, uint index) internal pure returns (int256 result) {
+    function _incentiveRatio(ActionIncentive memory config_, uint256 index) internal pure returns (int256 result) {
         result = config_.slot1.decodeInt(index * 32, 32) * int256(10 ** (18 - INCENTIVE_RATIO_DECIMALS));
     }
 
@@ -91,19 +99,22 @@ library ConfigIncentiveLib {
 
     /// @notice Stores a collateral ratio bound at the given index
     // solhint-disable-next-line explicit-types
-    function _setCollateralRatioUpperBounds(ActionIncentive memory config_, uint index, uint256 value) internal pure {
+    function _setCollateralRatioUpperBounds(ActionIncentive memory config_, uint256 index, uint256 value)
+        internal
+        pure
+    {
         config_.slot0 = config_.slot0.encodeUint(value / 10 ** (18 - COLLATERAL_RATIO_DECIMALS), index * 32, 32);
     }
 
     /// @notice Stored the collateral ratio bound count
     // solhint-disable-next-line explicit-types
-    function _setCollateralRatioBandCount(ActionIncentive memory config_, uint value) internal pure {
+    function _setCollateralRatioBandCount(ActionIncentive memory config_, uint256 value) internal pure {
         config_.slot0 = config_.slot0.encodeUint(value, 224, 8);
     }
 
     /// @notice Stores a incentive ratio at the given index
     // solhint-disable-next-line explicit-types
-    function _setIncentiveRatio(ActionIncentive memory config_, uint index, int256 value) internal pure {
+    function _setIncentiveRatio(ActionIncentive memory config_, uint256 index, int256 value) internal pure {
         config_.slot1 = config_.slot1.encodeInt(value / int256(10 ** (18 - INCENTIVE_RATIO_DECIMALS)), index * 32, 32);
     }
