+++ src/minter/library/Config_v1.sol
@@ -13,10 +13,10 @@ import {IMinter} from "src/interfaces/IMinter.sol";
 library Config_v1 {
     using ConfigIncentiveLib for ConfigIncentiveLib.ActionIncentive;
 
-    uint public constant MINT_PEGGED = 0; // solhint-disable-line explicit-types
-    uint public constant REDEEM_PEGGED = 1; // solhint-disable-line explicit-types
-    uint public constant MINT_LEVERAGED = 2; // solhint-disable-line explicit-types
-    uint public constant REDEEM_LEVERAGED = 3; // solhint-disable-line explicit-types
+    uint256 public constant MINT_PEGGED = 0; // solhint-disable-line explicit-types
+    uint256 public constant REDEEM_PEGGED = 1; // solhint-disable-line explicit-types
+    uint256 public constant MINT_LEVERAGED = 2; // solhint-disable-line explicit-types
+    uint256 public constant REDEEM_LEVERAGED = 3; // solhint-disable-line explicit-types
 
     /// @notice Checks a given incentive config for errors and returns it ready for storage
     /// @param name Label for error messages
@@ -26,27 +26,25 @@ library Config_v1 {
     /// false means it's a redeem pegged or mint leveraged config
     /// @return out the storage efficient config
     // slither-disable-next-line cyclomatic-complexity as this code is simple in what it tries to do, it's just that there are a few checks
-    function checkAndCopyBands(
-        string memory name,
-        IMinter.IncentiveConfig calldata config_,
-        bool disallowNotDiscount
-    ) internal pure returns (ConfigIncentiveLib.ActionIncentive memory out) {
+    function checkAndCopyBands(string memory name, IMinter.IncentiveConfig calldata config_, bool disallowNotDiscount)
+        internal
+        pure
+        returns (ConfigIncentiveLib.ActionIncentive memory out)
+    {
         // check the array sizes match
         if (config_.incentiveRatios.length < 1) {
             revert IMinter.TooFewIncentiveRatios(name, config_.incentiveRatios.length, 1);
         }
         if (config_.incentiveRatios.length != config_.collateralRatioBandUpperBounds.length + 1) {
             revert IMinter.CollateralRatioBoundsIncentivesLengthsMismatch(
-                name,
-                config_.collateralRatioBandUpperBounds.length,
-                config_.incentiveRatios.length
+                name, config_.collateralRatioBandUpperBounds.length, config_.incentiveRatios.length
             );
         }
         out = ConfigIncentiveLib.ActionIncentive(0, 0);
         uint256 prevUpperBound = 0;
-        uint iOut = 0; // solhint-disable-line explicit-types
+        uint256 iOut = 0; // solhint-disable-line explicit-types
         // solhint-disable-next-line explicit-types
-        for (uint i = 0; i < config_.incentiveRatios.length; i++) {
+        for (uint256 i = 0; i < config_.incentiveRatios.length; i++) {
             int256 incentiveRatio = ConfigIncentiveLib._incentiveRatioToStoragePrecision(config_.incentiveRatios[i]);
 
             // incentive ratios cannot be too precise for the storage schema
@@ -64,39 +62,27 @@ library Config_v1 {
                 // disallows, if they exist, must be at index 0
                 if (incentiveRatio == 1 ether && i != 0) {
                     revert IMinter.InvalidIncentiveRatioValue(
-                        name,
-                        i,
-                        config_.incentiveRatios[i],
-                        "disallow (1) must be at index 0"
+                        name, i, config_.incentiveRatios[i], "disallow (1) must be at index 0"
                     );
                 }
             } else {
                 // it's a redeem pegged or mint leveraged
                 // check against interval (-1, 1) i.e. some discount; to zero; to some fees
                 if (incentiveRatio <= -1 ether || incentiveRatio >= 1 ether) {
-                    revert IMinter.InvalidIncentiveRatioValue(
-                        name,
-                        i,
-                        config_.incentiveRatios[i],
-                        "must be in (-1, 1)"
-                    );
+                    revert IMinter.InvalidIncentiveRatioValue(name, i, config_.incentiveRatios[i], "must be in (-1, 1)");
                 }
             }
             // check collateral ratio upper bounds are strictly increasing and then copy
             uint256 currentUpperBound;
             if (i < config_.collateralRatioBandUpperBounds.length) {
-                currentUpperBound = ConfigIncentiveLib._collateralRatioToStoragePrecision(
-                    config_.collateralRatioBandUpperBounds[i]
-                );
+                currentUpperBound =
+                    ConfigIncentiveLib._collateralRatioToStoragePrecision(config_.collateralRatioBandUpperBounds[i]);
                 if (currentUpperBound != config_.collateralRatioBandUpperBounds[i]) {
                     revert IMinter.CollateralRatioBoundTooPrecise(name, config_.collateralRatioBandUpperBounds[i]);
                 }
                 if (i == 0 && currentUpperBound < 1 ether) {
                     revert IMinter.InvalidCollateralRatioBoundValue(
-                        name,
-                        currentUpperBound,
-                        i,
-                        "first boundary must be >= 1"
+                        name, currentUpperBound, i, "first boundary must be >= 1"
                     );
                 }
                 if (i > 0 && currentUpperBound <= 1 ether) {
@@ -129,9 +115,7 @@ library Config_v1 {
             }
             if (iOut >= ConfigIncentiveLib.MAX_BANDS) {
                 revert IMinter.TooManyIncentiveRatios(
-                    name,
-                    config_.incentiveRatios.length,
-                    config_.incentiveRatios.length - 1
+                    name, config_.incentiveRatios.length, config_.incentiveRatios.length - 1
                 );
             }
 
@@ -147,37 +131,31 @@ library Config_v1 {
         return out;
     }
 
-    function checkAndCopyIncentives(
-        IMinter.Config calldata config_,
-        ConfigIncentiveLib.ActionIncentive[4] storage out
-    ) external {
+    function checkAndCopyIncentives(IMinter.Config calldata config_, ConfigIncentiveLib.ActionIncentive[4] storage out)
+        external
+    {
         out[MINT_PEGGED] = Config_v1.checkAndCopyBands("mint pegged", config_.mintPeggedIncentiveConfig, true);
         out[REDEEM_PEGGED] = Config_v1.checkAndCopyBands("redeem pegged", config_.redeemPeggedIncentiveConfig, false);
-        out[MINT_LEVERAGED] = Config_v1.checkAndCopyBands(
-            "mint leveraged",
-            config_.mintLeveragedIncentiveConfig,
-            false
-        );
-        out[REDEEM_LEVERAGED] = Config_v1.checkAndCopyBands(
-            "redeem leveraged",
-            config_.redeemLeveragedIncentiveConfig,
-            true
-        );
+        out[MINT_LEVERAGED] = Config_v1.checkAndCopyBands("mint leveraged", config_.mintLeveragedIncentiveConfig, false);
+        out[REDEEM_LEVERAGED] =
+            Config_v1.checkAndCopyBands("redeem leveraged", config_.redeemLeveragedIncentiveConfig, true);
     }
 
     /// @notice Converts the compact storage format back to the full IncentiveConfig
     /// @param config_ The storage-efficient configuration to convert back
     /// @return out The user-friendly config structure
-    function copyBandsBack(
-        ConfigIncentiveLib.ActionIncentive memory config_
-    ) internal pure returns (IMinter.IncentiveConfig memory out) {
-        uint iOut = 0; // solhint-disable-line explicit-types
-        uint outBands = ConfigIncentiveLib._collateralRatioBandCount(config_); // solhint-disable-line explicit-types
-        uint outBounds = outBands - 1; // solhint-disable-line explicit-types
+    function copyBandsBack(ConfigIncentiveLib.ActionIncentive memory config_)
+        internal
+        pure
+        returns (IMinter.IncentiveConfig memory out)
+    {
+        uint256 iOut = 0; // solhint-disable-line explicit-types
+        uint256 outBands = ConfigIncentiveLib._collateralRatioBandCount(config_); // solhint-disable-line explicit-types
+        uint256 outBounds = outBands - 1; // solhint-disable-line explicit-types
         out.collateralRatioBandUpperBounds = new uint256[](outBounds);
         out.incentiveRatios = new int256[](outBands);
         // solhint-disable-next-line explicit-types
-        for (uint i = 0; i < outBounds; i++) {
+        for (uint256 i = 0; i < outBounds; i++) {
             uint256 ub = ConfigIncentiveLib._collateralRatioUpperBounds(config_, i);
             if (ub == 1 ether - 1) {
                 ub = 1 ether;
@@ -190,9 +168,11 @@ library Config_v1 {
         return out;
     }
 
-    function copyIncentivesBack(
-        ConfigIncentiveLib.ActionIncentive[4] memory config_
-    ) internal pure returns (IMinter.Config memory out) {
+    function copyIncentivesBack(ConfigIncentiveLib.ActionIncentive[4] memory config_)
+        internal
+        pure
+        returns (IMinter.Config memory out)
+    {
         out.mintPeggedIncentiveConfig = copyBandsBack(config_[MINT_PEGGED]);
         out.redeemPeggedIncentiveConfig = copyBandsBack(config_[REDEEM_PEGGED]);
         out.mintLeveragedIncentiveConfig = copyBandsBack(config_[MINT_LEVERAGED]);
